gui-web/requirements.txt
# setuptools
# python=3
# dcm2niix. For more information, see ./docs/installation.md and ./heuristics/bidsmap_template.yaml
astroid==2.2.5
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Cors==3.0.7
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.2
pandas==0.24.2
pydicom==1.2.2
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
ruamel.yaml==0.15.91
six==1.12.0
typed-ast==1.3.1
Werkzeug==0.15.2
wrapt==1.11.1
